+++ packages/amazon/src/serverGroup/details/scalingPolicy/upsert/step/StepPolicyAction.tsx
@@ -2,12 +2,12 @@ import * as React from 'react';
 
 import { NumberInput, ReactSelectInput } from '@spinnaker/core';
 
-import { IScalingPolicyAlarm, IStepAdjustment } from '../../../../../domain';
+import type { IScalingPolicyAlarm, IStepAdjustment } from '../../../../../domain';
 
 import './StepPolicyAction.less';
 
-type Operator = 'Add' | 'Remove' | 'Set to';
-type AdjustmentTypeView = 'instances' | 'percent of group';
+export type Operator = 'Add' | 'Remove' | 'Set to';
+export type AdjustmentTypeView = 'instances' | 'percent of group';
 
 export interface IStepPolicyActionProps {
   adjustmentType: AdjustmentTypeView;
@@ -33,51 +33,40 @@ export const StepPolicyAction = ({
   const hasEqualTo = alarm?.comparisonOperator.includes('Equal');
   const availableActions = ['Add', 'Remove', 'Set to'];
 
-  const [action, setAction] = React.useState<Operator>(operator);
-  const adjustmentTypeOptions = action === 'Set to' ? ['instances'] : ['instances', 'percent of group'];
+  const adjustmentTypeOptions = operator === 'Set to' ? ['instances'] : ['instances', 'percent of group'];
   const onActionChange = (val: Operator) => {
-    setAction(val);
     adjustmentTypeChanged(val, adjustmentType);
   };
 
-  const [adjustmentTypeView, setAdjustmentTypeView] = React.useState<AdjustmentTypeView>(adjustmentType);
   const onAdjustmentTypeChange = (type: AdjustmentTypeView) => {
-    setAdjustmentTypeView(type);
-    adjustmentTypeChanged(action, type);
+    adjustmentTypeChanged(operator, type);
   };
 
-  const [steps, setSteps] = React.useState<IStepAdjustment[]>(step?.stepAdjustments || stepAdjustments);
+  const steps = step?.stepAdjustments || stepAdjustments;
   const addStep = () => {
     const newStep = { scalingAdjustment: 1 } as IStepAdjustment;
     const newSteps = [...steps, newStep];
-    setSteps(newSteps);
     stepsChanged(newSteps);
   };
   const removeStep = (index: number) => {
     const newSteps = steps.filter((_s, i) => i !== index);
-    setSteps(newSteps);
     stepsChanged(newSteps);
   };
   const updateStep = (updatedStep: IStepAdjustment, index: number) => {
     const newSteps = [...steps];
     newSteps[index] = updatedStep;
-    setSteps(newSteps);
     stepsChanged(newSteps);
   };
-
-  React.useEffect(() => {
-    setSteps(step.stepAdjustments);
-  }, [step?.stepAdjustments]);
 
   return (
     <div className="StepPolicyAction row">
       {steps?.map((step: IStepAdjustment, index: number) => (
         <div key={`step-adjustment-${index}`} className="step-policy-row col-md-10 col-md-offset-1 horizontal middle">
           {Boolean(index) ? (
-            <span className="action-input sp-margin-xs-left">{action}</span>
+            <span className="action-input sp-margin-xs-left">{operator}</span>
           ) : (
             <ReactSelectInput
-              value={action}
+              value={operator}
               stringOptions={availableActions}
               onChange={(e) => onActionChange(e.target.value)}
               clearable={false}
@@ -91,14 +80,14 @@ export const StepPolicyAction = ({
             inputClassName="action-input"
           />
           {Boolean(index) ? (
-            <span className="sp-margin-xs-left">{adjustmentTypeView}</span>
+            <span className="sp-margin-xs-left">{adjustmentType}</span>
           ) : (
             <ReactSelectInput
-              value={adjustmentTypeView}
+              value={adjustmentType}
               stringOptions={adjustmentTypeOptions}
               onChange={(e) => onAdjustmentTypeChange(e.target.value)}
               clearable={false}
-              inputClassName="adjustment-type-input"
+              inputClassName="adjustment-type-input sp-margin-xs-left"
             />
           )}
           <span className="sp-margin-xs-xaxis">
@@ -150,7 +139,7 @@ export const StepPolicyAction = ({
       ))}
       <div className="row sp-margin-s">
         <div className="col-md-10 col-md-offset-1">
-          <button className="btn btn-block btn-sm add-new" onClick={addStep}>
+          <button type="button" className="btn btn-block btn-sm add-new" onClick={addStep}>
             <span className="glyphicon glyphicon-plus-sign"></span>
             Add step
           </button>