+++ packages/core/src/config/settings.ts
@@ -67,6 +67,15 @@ export interface INewApplicationDefaults {
   chaosMonkey?: boolean;
 }
 
+export interface IManagedDeliveryURLs {
+  gettingStarted: string;
+  root: string;
+  pinning: string;
+  resourceStatus: string;
+  markAsBad: string;
+  previewEnvironments?: string;
+}
+
 export interface ISpinnakerSettings {
   [key: string]: any;
 
@@ -111,9 +120,9 @@ export interface ISpinnakerSettings {
   gitSources: string[];
   hiddenStages: string[];
   managedDelivery?: {
-    gettingStartedUrl?: string;
     defaultManifest: string;
     manifestBasePath: string;
+    urls?: Partial<IManagedDeliveryURLs>;
   };
   maxPipelineAgeDays: number;
   newApplicationDefaults: INewApplicationDefaults;
@@ -141,6 +150,7 @@ export interface ISpinnakerSettings {
   triggerTypes: string[];
   useClassicFirewallLabels: boolean;
   kubernetesAdHocInfraWritesEnabled: boolean;
+  awsAdHocInfraWritesEnabled: boolean;
   changelogUrl: string;
 }
 
@@ -149,6 +159,8 @@ export const SETTINGS: ISpinnakerSettings = (window as any).spinnakerSettings ||
 // Make sure to set up some reasonable default settings fields so we do not have to keep checking if they exist everywhere
 SETTINGS.feature = SETTINGS.feature || {};
 SETTINGS.feature.roscoMode = SETTINGS.feature.roscoMode ?? true;
+SETTINGS.kubernetesAdHocInfraWritesEnabled = SETTINGS.kubernetesAdHocInfraWritesEnabled ?? true;
+SETTINGS.awsAdHocInfraWritesEnabled = SETTINGS.awsAdHocInfraWritesEnabled ?? true;
 SETTINGS.analytics = SETTINGS.analytics || {};
 SETTINGS.providers = SETTINGS.providers || {};
 SETTINGS.defaultTimeZone = SETTINGS.defaultTimeZone || 'America/Los_Angeles';