+++ packages/core/src/managed/graphql/graphql-sdk.ts
@@ -120,6 +120,7 @@ export interface MdConfig {
   updatedAt?: Maybe<Scalars['InstantTime']>;
   rawConfig?: Maybe<Scalars['String']>;
   processedConfig?: Maybe<Scalars['String']>;
+  previewEnvironmentsConfigured?: Maybe<Scalars['Boolean']>;
 }
 
 export interface MdConstraint {
@@ -142,6 +143,20 @@ export interface MdConstraintStatusPayload {
   version: Scalars['String'];
   reference: Scalars['String'];
   status: MdConstraintStatus;
+}
+
+export interface MdDeployLocation {
+  __typename?: 'MdDeployLocation';
+  account?: Maybe<Scalars['String']>;
+  region?: Maybe<Scalars['String']>;
+  sublocations?: Maybe<Array<Scalars['String']>>;
+}
+
+export interface MdDeployTarget {
+  __typename?: 'MdDeployTarget';
+  cloudProvider?: Maybe<Scalars['String']>;
+  location?: Maybe<MdDeployLocation>;
+  status?: Maybe<MdRolloutTargetStatus>;
 }
 
 export interface MdDismissNotificationPayload {
@@ -168,6 +183,16 @@ export interface MdEnvironmentState {
 }
 
 export type MdEventLevel = 'SUCCESS' | 'INFO' | 'WARNING' | 'ERROR';
+
+export interface MdExecutionSummary {
+  __typename?: 'MdExecutionSummary';
+  status: MdTaskStatus;
+  currentStage?: Maybe<MdStageDetail>;
+  stages?: Maybe<Array<MdStageDetail>>;
+  deployTargets?: Maybe<Array<MdDeployTarget>>;
+  completedDeployTargets?: Maybe<Array<MdDeployTarget>>;
+  error?: Maybe<Scalars['String']>;
+}
 
 export interface MdGitIntegration {
   __typename?: 'MdGitIntegration';
@@ -317,6 +342,7 @@ export interface MdResourceTask {
   __typename?: 'MdResourceTask';
   id: Scalars['String'];
   name: Scalars['String'];
+  summary?: Maybe<MdExecutionSummary>;
 }
 
 export interface MdRetryArtifactActionPayload {
@@ -327,6 +353,34 @@ export interface MdRetryArtifactActionPayload {
   actionId: Scalars['String'];
   actionType: MdActionType;
 }
+
+export type MdRolloutTargetStatus = 'NOT_STARTED' | 'RUNNING' | 'SUCCEEDED' | 'FAILED';
+
+export interface MdStageDetail {
+  __typename?: 'MdStageDetail';
+  id?: Maybe<Scalars['String']>;
+  type?: Maybe<Scalars['String']>;
+  name?: Maybe<Scalars['String']>;
+  startTime?: Maybe<Scalars['InstantTime']>;
+  endTime?: Maybe<Scalars['InstantTime']>;
+  status?: Maybe<MdTaskStatus>;
+  refId?: Maybe<Scalars['String']>;
+  requisiteStageRefIds?: Maybe<Array<Scalars['String']>>;
+}
+
+export type MdTaskStatus =
+  | 'NOT_STARTED'
+  | 'RUNNING'
+  | 'PAUSED'
+  | 'SUSPENDED'
+  | 'SUCCEEDED'
+  | 'FAILED_CONTINUE'
+  | 'TERMINAL'
+  | 'CANCELED'
+  | 'REDIRECT'
+  | 'STOPPED'
+  | 'BUFFERED'
+  | 'SKIPPED';
 
 export interface MdToggleResourceManagementPayload {
   id: Scalars['ID'];
@@ -492,6 +546,7 @@ export type FetchApplicationQueryVariables = Exact<{
 export type FetchApplicationQuery = { __typename?: 'Query' } & {
   application?: Maybe<
     { __typename?: 'MdApplication' } & Pick<MdApplication, 'id' | 'name' | 'account'> & {
+        config?: Maybe<{ __typename?: 'MdConfig' } & Pick<MdConfig, 'id' | 'previewEnvironmentsConfigured'>>;
         environments: Array<
           { __typename?: 'MdEnvironment' } & Pick<MdEnvironment, 'isDeleting'> & {
               state: { __typename?: 'MdEnvironmentState' } & Pick<MdEnvironmentState, 'id'> & {
@@ -695,7 +750,7 @@ export type FetchApplicationManagementDataQuery = { __typename?: 'Query' } & {
         gitIntegration?: Maybe<
           { __typename?: 'MdGitIntegration' } & Pick<
             MdGitIntegration,
-            'id' | 'repository' | 'branch' | 'isEnabled' | 'link'
+            'id' | 'repository' | 'branch' | 'isEnabled' | 'link' | 'manifestPath'
           >
         >;
       }
@@ -774,6 +829,12 @@ export type ImportDeliveryConfigMutationVariables = Exact<{
 }>;
 
 export type ImportDeliveryConfigMutation = { __typename?: 'Mutation' } & Pick<Mutation, 'importDeliveryConfig'>;
+
+export type ToggleResourceManagementMutationVariables = Exact<{
+  payload?: Maybe<MdToggleResourceManagementPayload>;
+}>;
+
+export type ToggleResourceManagementMutation = { __typename?: 'Mutation' } & Pick<Mutation, 'toggleResourceManagement'>;
 
 export const ActionDetailsFragmentDoc = gql`
   fragment actionDetails on MdAction {
@@ -878,6 +939,10 @@ export const FetchApplicationDocument = gql`
       id
       name
       account
+      config {
+        id
+        previewEnvironmentsConfigured
+      }
       environments {
         ...baseEnvironmentFields
         isDeleting
@@ -1309,6 +1374,7 @@ export const FetchApplicationManagementDataDocument = gql`
         branch
         isEnabled
         link
+        manifestPath
       }
     }
   }
@@ -1833,4 +1899,46 @@ export type ImportDeliveryConfigMutationResult = Apollo.MutationResult<ImportDel
 export type ImportDeliveryConfigMutationOptions = Apollo.BaseMutationOptions<
   ImportDeliveryConfigMutation,
   ImportDeliveryConfigMutationVariables
+>;
+export const ToggleResourceManagementDocument = gql`
+  mutation ToggleResourceManagement($payload: MdToggleResourceManagementPayload) {
+    toggleResourceManagement(payload: $payload)
+  }
+`;
+export type ToggleResourceManagementMutationFn = Apollo.MutationFunction<
+  ToggleResourceManagementMutation,
+  ToggleResourceManagementMutationVariables
+>;
+
+/**
+ * __useToggleResourceManagementMutation__
+ *
+ * To run a mutation, you first call `useToggleResourceManagementMutation` within a React component and pass it any options that fit your needs.
+ * When your component renders, `useToggleResourceManagementMutation` returns a tuple that includes:
+ * - A mutate function that you can call at any time to execute the mutation
+ * - An object with fields that represent the current status of the mutation's execution
+ *
+ * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
+ *
+ * @example
+ * const [toggleResourceManagementMutation, { data, loading, error }] = useToggleResourceManagementMutation({
+ *   variables: {
+ *      payload: // value for 'payload'
+ *   },
+ * });
+ */
+export function useToggleResourceManagementMutation(
+  baseOptions?: Apollo.MutationHookOptions<ToggleResourceManagementMutation, ToggleResourceManagementMutationVariables>,
+) {
+  const options = { ...defaultOptions, ...baseOptions };
+  return Apollo.useMutation<ToggleResourceManagementMutation, ToggleResourceManagementMutationVariables>(
+    ToggleResourceManagementDocument,
+    options,
+  );
+}
+export type ToggleResourceManagementMutationHookResult = ReturnType<typeof useToggleResourceManagementMutation>;
+export type ToggleResourceManagementMutationResult = Apollo.MutationResult<ToggleResourceManagementMutation>;
+export type ToggleResourceManagementMutationOptions = Apollo.BaseMutationOptions<
+  ToggleResourceManagementMutation,
+  ToggleResourceManagementMutationVariables
 >;